Update rtk_argb_lib to new architecture
- Using one lib for RtkArgbController and RtkBridgeController
