--- Controllers/RtkArgbApi/hid_dev_spti.h
@@ -1,64 +1,0 @@
-/*
- * BSD-3-Clause
- *
- * Copyright (c) 2022 - 2024, Realtek Semiconductor Corp. All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without modification,
- * are permitted provided that the following conditions are met:
- *
- *   * Redistributions of source code must retain the above copyright notice, this
- *     list of conditions and the following disclaimer.
- *
- *   * Redistributions in binary form must reproduce the above copyright notice,
- *     this list of conditions and the following disclaimer in the documentation
- *     and/or other materials provided with the distribution.
- *
- *   * Neither the name of the Realtek nor the names of its contributors may
- *     be used to endorse or promote products derived from this software without
- *     specific prior written permission.
- *
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
- * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS
- * OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
- * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN
- * IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-#ifndef __HID_DEV_SPTI_H__
-#define __HID_DEV_SPTI_H__
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-#ifdef CONFIG_HID
-#ifdef CONFIG_ARGB_DEV
-#define M_HID_DEV_SPTI		argb_hid_dev_spti
-#define M_HID_WRITE_UNLOCK	argb_hid_write_unlock
-#else
-#define M_HID_DEV_SPTI		hid_dev_spti
-#define M_HID_WRITE_UNLOCK	hid_write_unlock
-#endif
-#endif
-
-
-#ifdef CONFIG_HID
-#ifdef CONFIG_ARGB_DEV
-int argb_hid_dev_spti(hid_device *fd, uint8_t *cdb, uint8_t *buf, int buflen, uint8_t is_in, int timeout, uint8_t *sense);
-int argb_hid_write_unlock(hid_device *fd);
-#else
-int hid_dev_spti(hid_device *fd, uint8_t *cdb, uint8_t *buf, int buflen, uint8_t is_in, int timeout, uint8_t *sense);
-int hid_write_unlock(hid_device *fd);
-#endif
-#endif
-
-#ifdef __cplusplus
-}
-#endif
-#endif /* __HID_DEV_SPTI_H__ */
