+++ Controllers/RtkArgbController/rtk_argb_lib/include/api_ver.h
@@ -1,0 +1,46 @@
+/*
+ * BSD-3-Clause
+ *
+ * Copyright (c) 2022 - 2025, Realtek Semiconductor Corp. All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without modification,
+ * are permitted provided that the following conditions are met:
+ *
+ *   * Redistributions of source code must retain the above copyright notice, this
+ *     list of conditions and the following disclaimer.
+ *
+ *   * Redistributions in binary form must reproduce the above copyright notice,
+ *     this list of conditions and the following disclaimer in the documentation
+ *     and/or other materials provided with the distribution.
+ *
+ *   * Neither the name of the Realtek nor the names of its contributors may
+ *     be used to endorse or promote products derived from this software without
+ *     specific prior written permission.
+ *
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+ * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
+ * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS
+ * OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
+ * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN
+ * IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+#pragma once
+
+#define UTHSB_API_VER_MAIN1 1
+#define UTHSB_API_VER_MAIN2 2
+#define UTHSB_API_VER_SUB1 8
+#define UTHSB_API_VER_SUB2 50919
+
+
+// version number (string)
+#define TOSTRING2(arg) #arg
+#define TOSTRING(arg) TOSTRING2(arg)
+#define RELEASE_VER_STR     TOSTRING(UTHSB_API_VER_MAIN1) "." TOSTRING(UTHSB_API_VER_MAIN2) "." TOSTRING(UTHSB_API_VER_SUB1) "." TOSTRING(UTHSB_API_VER_SUB2)
+
+#pragma once
