+++ Controllers/RtkArgbController/rtk_argb_lib/include/argb_dev.h
@@ -1,0 +1,77 @@
+/*
+ * BSD-3-Clause
+ *
+ * Copyright (c) 2022 - 2024, Realtek Semiconductor Corp. All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without modification,
+ * are permitted provided that the following conditions are met:
+ *
+ *   * Redistributions of source code must retain the above copyright notice, this
+ *     list of conditions and the following disclaimer.
+ *
+ *   * Redistributions in binary form must reproduce the above copyright notice,
+ *     this list of conditions and the following disclaimer in the documentation
+ *     and/or other materials provided with the distribution.
+ *
+ *   * Neither the name of the Realtek nor the names of its contributors may
+ *     be used to endorse or promote products derived from this software without
+ *     specific prior written permission.
+ *
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+ * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
+ * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS
+ * OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
+ * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN
+ * IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+#ifndef _ARGB_DEV_H_
+#define _ARGB_DEV_H_
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+#include "bridge_dev.h"
+#define argb_device bridge_device
+
+//enum {
+//	ARGB_DEV_HID = 0,
+//	ARGB_DEV_SCSI,
+//	ARGB_DEV_USB
+//};
+
+struct argb_dev_info {
+	char *name;
+	char *path;
+	int intf_type;
+	struct argb_dev_info *next;
+};
+
+//struct argb_device {
+//	void *hdl;
+//	int type;
+//	struct bridge_hid_ch *chs;
+//};
+
+struct argb_dev_info *argb_dev_scan(struct usb_id *id, int id_num);
+
+void argb_dev_free_scan(struct argb_dev_info *devs);
+
+struct argb_device* argb_path_open(char* path, int type, struct bridge_hid_ch *extchs);
+
+struct argb_device *argb_dev_open(struct argb_dev_info *devs, const char *name, struct bridge_hid_ch *extchs);
+
+void argb_dev_close(struct argb_device *dev);
+
+void *argb_dev_get_handle(struct argb_device *dev);
+#ifdef __cplusplus
+}
+#endif
+
+#endif /* _ARGB_DEV_H_ */
