+++ OpenRGB.pro
@@ -377,8 +377,7 @@ win32:contains(QMAKE_TARGET.arch, x86_64) {
         -L"$$PWD/dependencies/hidapi-win/x64/" -lhidapi                                         \
         -L"$$PWD/dependencies/mbedtls-3.2.1/lib/x64/" -lmbedcrypto -lmbedtls -lmbedx509         \
         -L"$$PWD/dependencies/PawnIO/" -lPawnIOLib                                              \
-        -L"$$PWD/Controllers/RtkArgbApi/rtk_argb_lib/windows/x64/" -lrtk_argb_libx64            \
-        -L"$$PWD/Controllers/RtkBridgeApi/rtk_bridge_lib/windows/x64/" -lrtk_uthsb_libx64       \
+        -L"$$PWD/Controllers/RtkArgbController/rtk_argb_lib/lib/windows/x64/" -lrtk_argb_libx64 \
 }
 
 win32:contains(QMAKE_TARGET.arch, x86) {
@@ -391,8 +390,7 @@ win32:contains(QMAKE_TARGET.arch, x86) {
         -L"$$PWD/dependencies/libusb-1.0.27/VS2019/MS32/dll" -llibusb-1.0                       \
         -L"$$PWD/dependencies/hidapi-win/x86/" -lhidapi                                         \
         -L"$$PWD/dependencies/mbedtls-3.2.1/lib/x86/" -lmbedcrypto -lmbedtls -lmbedx509         \
-        -L"$$PWD/Controllers/RtkArgbApi/rtk_argb_lib/windows/x86/" -lrtk_argb_libx86            \
-        -L"$$PWD/Controllers/RtkBridgeApi/rtk_bridge_lib/windows/x86/" -lrtk_uthsb_libx86       \
+        -L"$$PWD/Controllers/RtkArgbController/rtk_argb_lib/lib/windows/x86/" -lrtk_argb_libx86 \
 }
 
 win32:DEFINES -=                                                                                \
@@ -508,9 +506,15 @@ contains(QMAKE_PLATFORM, linux) {
     packagesExist(hidapi-hidraw) {
         PKGCONFIG += hidapi-hidraw
 
-        LIBS +=                                                                                 \
-        -L$$PWD/Controllers/RtkArgbApi/rtk_argb_lib/linux/hidraw/ -l_rtk_argbx86_64             \
-        -L$$PWD/Controllers/RtkBridgeApi/rtk_bridge_lib/linux/hidraw/ -l_rtk_uthsbx86_64        \
+        contains(QMAKE_HOST.arch, x86_64) {
+            LIBS +=                                                                                 \
+            -L$$PWD/Controllers/RtkArgbController/rtk_argb_lib/lib/linux/hidraw/ -l_rtk_argbx86_64  \
+        }
+
+        contains(QMAKE_HOST.arch, aarch64) {
+            LIBS +=                                                                                 \
+            -L$$PWD/Controllers/RtkArgbController/rtk_argb_lib/lib/linux/hidraw/ -l_rtk_argbaarch64 \
+        }
 
         #---------------------------------------------------------------------------------------#
         # hidapi-hidraw >= 0.10.1 supports USAGE/USAGE_PAGE                                     #
@@ -523,14 +527,27 @@ contains(QMAKE_PLATFORM, linux) {
     } else {
         packagesExist(hidapi-libusb) {
             PKGCONFIG += hidapi-libusb
-            LIBS +=                                                                             \
-            -L$$PWD/Controllers/RtkArgbApi/rtk_argb_lib/linux/libusb/ -l_rtk_argbx86_64         \
-            -L$$PWD/Controllers/RtkBridgeApi/rtk_bridge_lib/linux/libusb/ -l_rtk_uthsbx86_64    \
+
+            contains(QMAKE_HOST.arch, x86_64) {
+                LIBS +=                                                                                 \
+                -L$$PWD/Controllers/RtkArgbController/rtk_argb_lib/lib/linux/libusb/ -l_rtk_argbx86_64  \
+            }
+
+            contains(QMAKE_HOST.arch, aarch64) {
+                LIBS +=                                                                                 \
+                -L$$PWD/Controllers/RtkArgbController/rtk_argb_lib/lib/linux/libusb/ -l_rtk_argbaarch64 \
+            }
         } else {
             PKGCONFIG += hidapi
-            LIBS +=                                                                             \
-            -L$$PWD/Controllers/RtkArgbApi/rtk_argb_lib/linux/libusb/ -l_rtk_argbx86_64         \
-            -L$$PWD/Controllers/RtkBridgeApi/rtk_bridge_lib/linux/libusb/ -l_rtk_uthsbx86_64    \
+            contains(QMAKE_HOST.arch, x86_64) {
+                LIBS +=                                                                                 \
+                -L$$PWD/Controllers/RtkArgbController/rtk_argb_lib/lib/linux/libusb/ -l_rtk_argbx86_64  \
+            }
+
+            contains(QMAKE_HOST.arch, aarch64) {
+                LIBS +=                                                                                 \
+                -L$$PWD/Controllers/RtkArgbController/rtk_argb_lib/lib/linux/libusb/ -l_rtk_argbaarch64 \
+            }
         }
     }
 
@@ -785,8 +802,7 @@ macx:contains(QMAKE_HOST.arch, arm64) {
 
     LIBS +=                                                                                     \
     -L/opt/homebrew/lib                                                                         \
-    -L$$PWD/Controllers/RtkArgbApi/rtk_argb_lib/mac/arm64/ -l_rtk_argbarm64                     \
-    -L$$PWD/Controllers/RtkBridgeApi/rtk_bridge_lib/mac/arm64/ -l_rtk_uthsbarm64                \
+    -L$$PWD/Controllers/RtkArgbController/rtk_argb_lib/lib/mac/arm64/ -l_rtk_argbarm64          \
 }
 
 #-----------------------------------------------------------------------------------------------#
@@ -816,8 +832,7 @@ macx:contains(QMAKE_HOST.arch, x86_64) {
     LIBS +=                                                                                     \
     -L/usr/local/lib                                                                            \
     -L/usr/local/homebrew/lib                                                                   \
-    -L$$PWD/Controllers/RtkArgbApi/rtk_argb_lib/mac/x86_64/ -l_rtk_argbx86_64                   \
-    -L$$PWD/Controllers/RtkBridgeApi/rtk_bridge_lib/mac/x86_64/ -l_rtk_uthsbx86_64              \
+    -L$$PWD/Controllers/RtkArgbController/rtk_argb_lib/lib/mac/x86_64/ -l_rtk_argbx86_64        \
 
     DEFINES +=                                                                                  \
     _MACOSX_X86_X64                                                                             \
